% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLGCP_fixed.R
\name{sim_lgcp_multi_fixed}
\alias{sim_lgcp_multi_fixed}
\title{Simulation of semi-parametric multivariate log Gaussian Cox processes}
\usage{
sim_lgcp_multi_fixed(
  basecov,
  covariate,
  betas,
  alphas,
  xis,
  sigmas,
  phis,
  n.window,
  n.points,
  beta0s = NULL
)
}
\arguments{
\item{basecov}{Background intensity rho_0.}

\item{covariate}{Optional. A simulated covariate. The covariate must be a matrix.}

\item{betas}{A matrix with covariates.}

\item{alphas}{Alpha parameters. Must be a matrix, where the number of rows correspond
to the number of components in the LGCP. The number of columns correspond to the number
of common latent field.}

\item{xis}{Correlation scale parameters for each common random field. The correlation functions
for the common latent fields are exponential.}

\item{sigmas}{Sigma parameters. The number of sigma parameters must correspond
to the number of components in the LGCP.}

\item{phis}{Correlation scale parameters for each type-specific random field. The correlation functions
for the type-specific random fields are exponential.}

\item{n.window}{window size.}

\item{n.points}{Expected number of point for each component in the LGCP.
The length of n.points must correspond to the number of components.}

\item{beta0s}{Intercepts. The length of beta0s must correspond to the number of components
in the LGCP.}
}
\value{
Multivariate LGCP
}
\description{
Simulation of semi-parametric multivariate log Gaussian Cox processes.
}
\references{
Hessellund, K. B., Xu, G., Guan, Y. and Waagepetersen, R. (2020)
Second order semi-parametric inference for multivariate log Gaussian Cox processes.

This code has been updated on work done by Hessellund et al. (2022)'s sim-lgcp-multi code found in  https://github.com/kristianhessellund/Multilogreg.git
This version is able to address any given window and multiple covariates. 
(whereas Hessellund et al. (2022) used a square window and a single covariate)
}
\author{
Kristian Bjørn Hessellund, Ganggang Xu, Yongtao Guan and Rasmus Waagepetersen.
}
